vedic-agent-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pandas==2.2.2
numpy==1.26.4
pydantic==2.8.2
openai==1.30.1
python-dotenv==1.0.1
httpx==0.27.2  # Explicitly add httpx for compatibility
pymongo==4.8.0